// contracts/generated/typescript/models/EntityCatalogItem.ts
export interface EntityCatalogItem {
    /**
     *
     * @type {string}
     * @memberof EntityCatalogItem
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof EntityCatalogItem
     */
    wikiId: string;
    /**
     *
     * @type {string}
     * @memberof EntityCatalogItem
     */
    canonicalKey: string;
    /**
     *
     * @type {EntityCatalogItemStatusEnum}
     * @memberof EntityCatalogItem
     */
    status: EntityCatalogItemStatusEnum;
    /**
     *
     * @type {string}
     * @memberof EntityCatalogItem
     */
    mergedIntoEntityId?: string;
    /**
     *
     * @type {EntityTypeSummary}
     * @memberof EntityCatalogItem
     */
    entityType: EntityTypeSummary;
    /**
     *
     * @type {string}
     * @memberof EntityCatalogItem
     */
    displayLabel: string;
    /**
     *
     * @type {string}
     * @memberof EntityCatalogItem
     */
    displayLanguage: string;
    /**
     *
     * @type {string}
     * @memberof EntityCatalogItem
     */
    description: string;
    /**
     *
     * @type {number}
     * @memberof EntityCatalogItem
     */
    labelCount: number;
    /**
     *
     * @type {number}
     * @memberof EntityCatalogItem
     */
    aliasCount: number;
    /**
     *
     * @type {number}
     * @memberof EntityCatalogItem
     */
    claimCount: number;
    /**
     *
     * @type {number}
     * @memberof EntityCatalogItem
     */
    pageCount: number;
    /**
     *
     * @type {Date}
     * @memberof EntityCatalogItem
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof EntityCatalogItem
     */
    updatedAt: Date;
}


/**
 * @export
 */
export const EntityCatalogItemStatusEnum = {
    Active: 'active',
    Merged: 'merged'
} as const;
export type EntityCatalogItemStatusEnum = typeof EntityCatalogItemStatusEnum[keyof typeof EntityCatalogItemStatusEnum];