// contracts/generated/typescript/models/EntityMergeResult.ts
export interface EntityMergeResult {
    /**
     *
     * @type {string}
     * @memberof EntityMergeResult
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof EntityMergeResult
     */
    sourceEntityId: string;
    /**
     *
     * @type {string}
     * @memberof EntityMergeResult
     */
    targetEntityId: string;
    /**
     *
     * @type {string}
     * @memberof EntityMergeResult
     */
    actorId: string;
    /**
     *
     * @type {EntityMergeResultStatusEnum}
     * @memberof EntityMergeResult
     */
    status: EntityMergeResultStatusEnum;
    /**
     *
     * @type {string}
     * @memberof EntityMergeResult
     */
    reason: string;
    /**
     *
     * @type {Date}
     * @memberof EntityMergeResult
     */
    createdAt: Date;
    /**
     *
     * @type {boolean}
     * @memberof EntityMergeResult
     */
    idempotent: boolean;
    /**
     *
     * @type {Array<EntityMergeLabelMapping>}
     * @memberof EntityMergeResult
     */
    labelMappings: Array<EntityMergeLabelMapping>;
    /**
     *
     * @type {Array<EntityMergeClaimMapping>}
     * @memberof EntityMergeResult
     */
    claimMappings: Array<EntityMergeClaimMapping>;
}


/**
 * @export
 */
export const EntityMergeResultStatusEnum = {
    Applied: 'applied',
    RolledBack: 'rolled_back'
} as const;
export type EntityMergeResultStatusEnum = typeof EntityMergeResultStatusEnum[keyof typeof EntityMergeResultStatusEnum];